### 2016-07-10 #36 @whedon
Submission: gwdegree: A Shiny App to Aid Interpretation of Geometrically-Weighted Degree Estimates in Exponential Random Graph Models
**Submitting author:** @michaellevy (<a href="http://orcid.org/0000-0002-4188-2527">Michael Levy</a>)
**Repository:** <a href="https://github.com/michaellevy/gwdegree" target ="_blank">https://github.com/michaellevy/gwdegree</a>
**Version:** v0.1.1
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.57495" target ="_blank">http://dx.doi.org/10.5281/zenodo.57495</a>
