### 2016-07-11 #36 @amoeba
Grade: 2 (Minor revisions)

General comments:

The software is easy to use and I was pleased to find a high level of attention to detail on the design of the interactive tool and the accompanying writing. The software works and works well. I can see clear research application.

Specific notes requiring action:

- **License:** Repository does contain a plain-text `LICENSE` file but the file does not contain the text of an OSI-approved license. The `DESCRIPTION` file does, however. @arfon How is JOSS handling this case, which seems to apply to submissions of R packages that must go through submission to CRAN?
- **Version:** The GitHub repository does not have any releases. Submitted version is 0.1.0, CRAN version is 0.1.1. @arfon do we need to get the JOSS version in line with the CRAN version? I tested the CRAN-submitted 0.1.1 version.
- **Installation:**: While the software does install as described in the documentation (README.md in the repo), the information was hard to find. Suggest breaking the README into sections, including a section for 'Installation' or at least putting the installation information in an easier to find location.
- **Performance:** No specific performance claims were made, and I find the application performs adequately and I would expect the software to have acceptable performance on other computers.
- **Example usage:** The software is an interactive tool and I found the tool contained in-context instruction on usage.
- **Functionality documentation:** See above.
- **Community guidelines:**: No on (1), (2), and (3). Suggest adding these sections to README.
- **Paper**:
  - Paper contains "Use" instructions, which — according to my maybe-wrong interpretation of — the [JOSS Reviewer Guidelines](http://joss.theoj.org/about#reviewer_guidelines) forbid. Suggest removing the "Use" paragraph from the paper.

Comments on the README.me in the repo:

- README references 'tabs' but if the user starts with the readme, they haven't seen the app (and its tabs). Consider explaining or showing the app before referencing the tabs.
- The link "There is a working version online", 404s. Suggest fixing link.
- As noted specifically above, I would suggest breaking the README up into a few major sections, covering installation, usage, a high-level overview of the software, and the community guidelines sections. The writing you have is already very good.

Please let me know if I can explain anything in more detail. Thanks!
