### 2016-07-11 dbdbc47
Minor revisions for JOSS
