### 2016-07-11 #36 @michaellevy
Thanks for the quick review, @amoeba! I think I've addressed all these things in michaellevy/gwdegree@dbdbc47. A couple points:

- v0.1.0 is what I originally submitted, but v0.1.1 (current on CRAN and GitHub) is much prefered. Recognizing the journal wants to review just one version, I hope this can be accommodated.

- I left the LICENSE file as is, per CRAN's protocol; not sure if it would bother anyone to include the text of the license there. Happy to test this on my next update.
