### 2016-07-11 #36 @amoeba
Grade: 1 (Accept)

I reviewed your changes and they address all of my concerns. Thanks!

I subsequently checked off the remaining items on the Reviewer questions checklist. In #37, it seems like JOSS may be okay with R packages that are on CRAN following the CRAN guidelines, which you have done. I've checked off the corresponding list item. It really does look like CRAN will reject a submission that includes the text of a license in a `LICENSE` file.

@arfon I reviewed v 0.1.1, and this submission is for 0.1.0. Can we just change this submission to be for 0.1.1 and be done?
