### 2016-07-12 #36 @arfon
> - I left the LICENSE file as is, per CRAN's protocol; not sure if it would bother anyone to include the text of the license there. Happy to test this on my next update.

This is OK I think @michaellevy

Thanks for the rapid review @amoeba!

> @arfon I reviewed v 0.1.1, and this submission is for 0.1.0. Can we just change this submission to be for 0.1.1 and be done?

Sounds good to me.

@michaellevy - I think the final thing here is to make an archive of the software in Zenodo/figshare/other similar service and update this thread with the archive DOI.
