### 2016-07-12 #36 @michaellevy
Great! Here is the archive: [![DOI](https://zenodo.org/badge/23239/michaellevy/gwdegree.svg)](https://zenodo.org/badge/latestdoi/23239/michaellevy/gwdegree)

Thanks @arfon and @amoeba, this whole process has been smooth, fast, and edifying.
