### 2016-07-12 #36 @arfon
@michaellevy - this is now accepted into JOSS! Your DOI will be http://dx.doi.org/10.21105/joss.00036 . The Crossref processing queues are a little backed up right now but the DOI should be live in the next couple of hours ⚡ 🚀 💥
